Basic event functionality implemented, still not finished.
Eventualities such as edited events are still not contemplated, and are TODO.

Event deletions, manual event starts and edits to discord events are not detected.

background task that checks if the event has started unconsistently throws Exception due to ts.ip being detected as a url.

on_message event is a test method that will be deleted later.